+++ tests/model/test_reasoning_mistral.py
@@ -1,3 +1,4 @@
+from pydantic_core import to_json
 from test_helpers.utils import skip_if_no_mistral
 
 from inspect_ai import Task, eval
@@ -26,4 +27,4 @@ def test_mistral_reasoning():
     )[0]
     assert log.status == "success"
     assert log.samples
-    assert isinstance(log.samples[0].messages[-1].content[0], ContentReasoning)
+    assert isinstance(log.samples[0].messages[1].content[0], ContentReasoning)